Refactor: MapContext 추가 지도 관련 함수 MapContext로 이동
